% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{bundle}
\alias{bundle}
\title{Package a service for deployment}
\usage{
bundle(nest, file = stop("file must be specified"))
}
\arguments{
\item{A}{character. The file to save the}
}
\description{
This function will package a service for deployment
}

