% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{data_selection}
\alias{data_selection}
\title{Select Data with Early Stopping Based on Standard Errors}
\usage{
data_selection(X, data_size, min_SE = NULL, log = FALSE, period = 1000)
}
\arguments{
\item{X}{Covariate matrix (p x n)}

\item{data_size}{Number of data points to select}

\item{min_SE}{Target standard errors for convergence}

\item{log}{Whether to print progress}

\item{period}{If log = TRUE, print progress every period}
}
\value{
A vector of selected data point indices
}
\description{
Performs greedy data selection with an early stopping criterion when standard errors converge.
}
