% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{data_selection_fixed}
\alias{data_selection_fixed}
\title{Select Informative Data Points for Modeling}
\usage{
data_selection_fixed(X, data_size, log = FALSE, period = 1000)
}
\arguments{
\item{X}{Covariate matrix (p x n)}

\item{data_size}{Number of data points to select}

\item{log}{Whether to print progress every 1000 iterations}

\item{period}{If log = TRUE, print progress every period}
}
\value{
A vector of selected data point indices
}
\description{
Selects a subset of data points to minimize standard error using greedy optimization.
}
