% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg.R
\name{egg}
\alias{egg}
\title{Create a custom endpoint}
\usage{
egg(model, predict = NULL, ...)
}
\description{
This creates an API endpoint. Model specific preprocessing and postprocessing
are added for several model types
}

