% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{expand_covariate_matrix}
\alias{expand_covariate_matrix}
\title{Expand Covariate Matrix Based on Cell Type Proportions}
\usage{
expand_covariate_matrix(
  X,
  lambda,
  family = "gaussian",
  keep_coef = matrix(TRUE, ncol(X), ncol(lambda)),
  lib_size = rep(1, nrow(X)),
  min_reads_per_1000 = 1,
  max_noise_sd = 1
)
}
\arguments{
\item{X}{Covariate matrix of dimension (n times p)}

\item{lambda}{Matrix of cell type proportions (n x k)}

\item{family}{Model family: "poisson", "negative binomial", or "gaussian"}

\item{keep_coef}{Matrix (p by k) indicating which coefficients to keep in case one knows what to remove a priori(default: all TRUEs)}

\item{lib_size}{Library sizes for each sample (default: 1 for all)}

\item{min_reads_per_1000}{If family == "poisson" or "negative binomial" Minimum reads per 1000 for robust scaling (default: 1)}

\item{max_noise_sd}{If family == "gaussian" maximum standard deviation of the error term to be used for robust scaling (default: 1)}
}
\value{
A scaled and expanded covariate matrix
}
\description{
Expands the design matrix \code{X} using cell type proportions \code{lambda},
applying scaling for sequencing depth and removing low-frequency features.
}
