% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GeneSetDb-methods.R,
%   R/SparrowResult-methods.R
\name{geneSetCollectionURLfunction}
\alias{geneSetCollectionURLfunction}
\alias{geneSetCollectionURLfunction<-}
\alias{geneSetCollectionURLfunction,GeneSetDb-method}
\alias{geneSetCollectionURLfunction<-,GeneSetDb-method}
\alias{geneSetCollectionURLfunction,SparrowResult-method}
\title{Get/set the gene set collection url function for a geneset collection}
\usage{
geneSetCollectionURLfunction(x, i, ...)

geneSetCollectionURLfunction(x, i) <- value

\S4method{geneSetCollectionURLfunction}{GeneSetDb}(x, i, ...)

\S4method{geneSetCollectionURLfunction}{GeneSetDb}(x, i) <- value

\S4method{geneSetCollectionURLfunction}{SparrowResult}(x, i, ...)
}
\description{
Reference \code{\link[=collectionMetadata]{collectionMetadata()}} for more info.
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetDb}: returns the gene set collection
url function from a GeneSetDb

\item \code{GeneSetDb}: sets the gene set collection url
function for a \code{GeneSetDb : Collection} combination.

\item \code{SparrowResult}: return the url function from a
\code{SparrowResult} object.
}}

