% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_endpoint.R
\name{lm_endpoint}
\alias{lm_endpoint}
\title{Create an endpoint from a \code{\link{lm}} object}
\usage{
lm_endpoint(model, predict = NULL, ...)
}
\arguments{
\item{model}{A model object. The model from which to build the endpoint}

\item{predict}{A predict function. The function to use to predict from the model}

\item{...}{additional arguments passed to predict}
}
\description{
Create an endpoint from a \code{\link{lm}} object
}

