% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{londonYear}
\alias{londonYear}
\title{Annual Temperature Data for London from 1800 - 2012}
\format{
A data frame with 213 rows and 2 variables:
\describe{
\item{year}{year, in date format}
\item{avgTemp}{average temperature of London, in Celsius}
}
}
\source{
\url{https://www.kaggle.com/berkeleyearth/climate-change-earth-surface-temperature-data}
}
\usage{
londonYear
}
\description{
A dataset that consists of the annual average temperature for
London, England starting in 1800 and ending in 2012.
}
\keyword{datasets}
