% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{min_effect_optimizer}
\alias{min_effect_optimizer}
\title{Compute Minimum Detectable Effect Size}
\usage{
min_effect_optimizer(SE, pow, alpha, acc = 0.001)
}
\arguments{
\item{SE}{Standard error}

\item{pow}{Desired power}

\item{alpha}{Type I error rate}

\item{acc}{Accuracy of search}
}
\description{
Calculates the minimum detectable effect size given standard error and desired power.
}
\keyword{internal}
