% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg.R
\name{model_summary}
\alias{model_summary}
\title{Returns a summary of the model}
\usage{
model_summary(egg)
}
\description{
This function returns the summary of the mode
for an endpoint as a character. Useful for the
webapp and displaying output
}

