% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{nest}
\alias{nest}
\title{Create a new service}
\usage{
nest(name = NULL, ...)
}
\arguments{
\item{name}{A character. The name of the service}

\item{rport}{An integer. The port for the R server}

\item{rurl}{A character. The URL for the Rserve server}

\item{pyport}{An interger. The port for the Flask server}
}
\description{
Create a new service
}

