% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/service.R
\name{package}
\alias{package}
\title{Bundle a service}
\usage{
package(service, file = stop("file must be specified"))

package(service, file = stop("file must be specified"))
}
\arguments{
\item{A}{character. The file to save the}
}
\description{
Bundles a service for deployment

This function will package a service for deployment
}

