% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{power_calc}
\alias{power_calc}
\title{Compute Statistical Power}
\usage{
power_calc(SE, effect, alpha = 0.05)
}
\arguments{
\item{SE}{Standard error}

\item{effect}{Effect size}

\item{alpha}{Type I error rate}
}
\description{
Calculates the power to detect an effect given a standard error and significance level.
}
\keyword{internal}
