% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{power_integral}
\alias{power_integral}
\title{Average Power Over a Range of Effect Sizes}
\usage{
power_integral(SE, effect_min, effect_max, alpha)
}
\arguments{
\item{SE}{Standard error}

\item{effect_min}{Minimum effect size}

\item{effect_max}{Maximum effect size}

\item{alpha}{Type I error rate}
}
\description{
Computes the average power over a uniform distribution of effect sizes.
}
\keyword{internal}
