% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{predict.endpoint}
\alias{predict.endpoint}
\title{Predict}
\usage{
\method{predict}{endpoint}(endpoint, newdata = NULL)
}
\arguments{
\item{endpoint}{An endpoint object}

\item{newdata}{A JSON object for input}
}
\description{
Predict from an endpoint
}
\details{
This assumes that the incoming data
}

