% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg.R
\name{predict_egg}
\alias{predict_egg}
\title{Predict}
\usage{
predict_egg(egg, newdata = NULL)
}
\arguments{
\item{newdata}{A JSON object for input}

\item{endpoint}{An endpoint object}
}
\description{
Predict from an egg
}

