% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-parquet.R
\name{read_parquet}
\alias{read_parquet}
\title{Read in data from Parquet into an R data frame via \code{reticulate}}
\usage{
read_parquet(path, columns = NULL)
}
\arguments{
\item{path}{path to Parquet file}

\item{columns}{optionally a character vector of columns to read (vs all the columns)}
}
\description{
Read in data from Parquet into an R data frame via \code{reticulate}
}
