% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{SE_optimizer}
\alias{SE_optimizer}
\title{Compute Minimum Standard Error for Given Power and Effect}
\usage{
SE_optimizer(min_effect, pow, alpha = 0.05, acc = 0.001)
}
\arguments{
\item{min_effect}{Minimum effect size}

\item{pow}{Desired power (e.g., 0.8 or 0.999)}

\item{alpha}{Type I error rate}

\item{acc}{Accuracy of the search}
}
\description{
Calculates the minimum standard error needed to detect a given effect with a specified power.
}
\keyword{internal}
