% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{service}
\alias{service}
\title{Create a new service}
\usage{
service(name = NULL, rport = 6311, rurl = "localhost", pyport = 5000)
}
\arguments{
\item{name}{A character. The name of the service}

\item{rport}{An integer. The port for the R server}

\item{rurl}{A character. The URL for the Rserve server}

\item{pyport}{An interger. The port for the Flask server}
}
\description{
Create a new service
}

