% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate gene expression data}
\usage{
simulate_data(
  n,
  nct,
  effect_range,
  intercept_range,
  min_effect = 0,
  library_size = 500,
  spot_ct = min(2, nct),
  p = 6,
  num_null = 2,
  prob_ct = NULL,
  family = "poisson",
  dispersion = 1
)
}
\arguments{
\item{n}{Number of data points}

\item{nct}{Number of cell types}

\item{effect_range}{Range of effect sizes (effect sizes drawn uniformly on this interval)}

\item{intercept_range}{Range for intercepts (intercept sizes drawn uniformly on this interval)}

\item{min_effect}{Minimum absolute value of effect sizes.}

\item{library_size}{Number of transcripts per cell/spot}

\item{spot_ct}{Number of cell types per spot}

\item{p}{Number of covariates}

\item{num_null}{Number of null coefficients to include}

\item{prob_ct}{Optional probabilities for cell type sampling}

\item{family}{The data generating distribution. One of poisson, negative binomial,binomial, and gaussian}

\item{dispersion}{Dispersion parameter for NB or Gaussian family. Size parameter for NB and sd for Gaussian. Default: 1}
}
\value{
A list with simulated y, X, lambda, beta, null_beta, and CT
}
\description{
This function simulates gene expression data using a spot-based model
}
