% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparrow-package.R
\docType{package}
\name{sparrow}
\alias{sparrow}
\alias{sparrow-package}
\title{Temporary Shorcut For Reading Arrow/Parquet Bit Into R via 'reticulate'}
\description{
Work is being done to make Parquet/Arrow a first-class R citizen
but -- until then -- I don't always want a Drill server round trip just
to read in some data and same goes for firing up a Spark instance (srsly).
So, this is a quick hack until the R packages are done.
}
\author{
Bob Rudis (bob@rud.is)
}
