% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{sparrow_methods}
\alias{sparrow_methods}
\title{Lists the supported GSEA methods by sparrow}
\usage{
sparrow_methods()
}
\value{
a character vector of GSEA names, or a list of metadata for each
method.
}
\description{
Lists the supported GSEA methods by sparrow
}
\examples{
sparrow_methods()
}
