% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sparrowhawk}
\alias{sparrowhawk}
\title{Provides an interval estimate for a population mean with a sample mean.}
\usage{
sparrowhawk(sampleMean, size, sdev, interval)
}
\arguments{
\item{sampleMean}{A number representing the mean of a sample.}

\item{size}{A number representing the number of values in a sample.}

\item{sdev}{A number representing the standard deviation of a sample.}

\item{interval}{A number representing the confidence interval estimate of a sample.}
}
\value{
Returns the degrees of freedom based on the \code{size}, the critical value based on the degrees of freedom and the \code{interval}, the standard error of the mean based on the \code{sdev} and the \code{size}, and the confidence interval based on the \code{sampleMean}, the critical values, and the standard error of the mean.
}
\description{
Provides an interval estimate for a population mean with a sample mean.
}
\examples{
sparrowhawk(200, 50, 20, .95)
}
