% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sparrowhype}
\alias{sparrowhype}
\title{Determines if the null hypothesis is rejected or if it is failed to be rejected.}
\usage{
sparrowhype(popMean, sampleMean, sdev, slvl, size)
}
\arguments{
\item{popMean}{A number representing the mean of a population.}

\item{sampleMean}{A number representing the mean of a sample.}

\item{sdev}{A number representing the standard deviation of a sample.}

\item{slvl}{A number representing the level of significance.}

\item{size}{A number representing the number of values in a sample.}
}
\value{
Returns the null and alternative hypotheses using the \code{popMean}, the confidence interval using the \code{slvl}, the critical value using the \code{size} and the confidence interval, the Z-Score using the \code{sampleMean}, the \code{popMean}, the \code{sdev}, and the \code{size}, and the P-Value using the Z-Score.
}
\description{
Determines if the null hypothesis is rejected or if it is failed to be rejected.
}
\examples{
sparrowhype(200, 203, 25, .05, 140)
sparrowhype(50, 49, 2, .01, 100)
}
