% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparrowResult-methods.R
\name{geneSetDb}
\alias{geneSetDb}
\title{Fetches the GeneSetDb from SparrowResult}
\usage{
geneSetDb(x)
}
\arguments{
\item{x}{\code{SparrowResult}}
}
\value{
The \code{GeneSetDb}
}
\description{
Fetches the GeneSetDb from SparrowResult
}
\examples{
vm <- exampleExpressionSet(do.voom=TRUE)
gdb <- exampleGeneSetDb()
mg <- seas(vm, gdb, design = vm$design, contrast = 'tumor')
geneSetDb(mg)
}
