% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{species_info}
\alias{species_info}
\title{Match a species query to the regularized species info.}
\usage{
species_info(query = NULL, ...)
}
\arguments{
\item{query}{the species name to lookup, if \code{NULL} (default), returns the
internal species info table}
}
\description{
Match a species query to the regularized species info.
}
\examples{
species_info()
species_info("human")
}
