% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Create an interactive volcano plot}
\usage{
volcanoPlot(
  x,
  stats = "dge",
  xaxis = "logFC",
  yaxis = "pval",
  idx,
  xtfrm = base::identity,
  ytfrm = function(vals) -log10(vals),
  xlab = xaxis,
  ylab = sprintf("-log10(\%s)", yaxis),
  highlight = NULL,
  horiz_line = c(padj = 0.1),
  xhex = NULL,
  yhex = NULL,
  width = NULL,
  height = NULL,
  shiny_source = "mgvolcano",
  ggtheme = theme_bw(),
  ...
)
}
\arguments{
\item{highlight}{A vector of featureIds to highlight, or a GeneSetDb
that we can extract the featureIds from for this purpose.}

\item{horiz_line}{A (optionally named) number vecor (length 1) that indicates
where a line should be drawn across the volcano plot. This is usually done
to signify statistical significance. When the number is "named", this
indicates that you want to find an approximation of the values plotted
on y based on some transformation of the values that is the named column
of x (like "padj"). The default value \code{c(padj = 0.10)} indicates you
want to draw a line at approximately where the adjust pvalue of 0.10 is
on the y-axis, which is the \emph{nominal} pvalues.}

\item{xhex}{The raw \code{.xv} (not \code{xtfrm(.xv)}) value that acts
as a threshold such that values less than this will be hexbinned.}

\item{yhex}{the \code{.yvt} value threshold. Vaues less than this will
be hexbinned.}
}
\description{
Convenience function to create volcano plots from results generated within
this package. This is mostly used by \code{{sparrow.shiny}}.
}
\examples{
mg <- exampleSparrowResult()
volcanoPlot(mg)
volcanoPlot(mg, xhex=1, yhex=0.05)
}
