% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate gene expression data}
\usage{
simulate_data(
  n,
  nct,
  effect_scale,
  intercept_scale,
  library_size,
  spot_ct,
  p,
  num_null = 3,
  prob_ct = NULL,
  family = "poisson",
  dispersion = NULL
)
}
\arguments{
\item{n}{Number of data points}

\item{nct}{Number of cell types}

\item{effect_scale}{Scale for effect sizes}

\item{intercept_scale}{Scale for intercepts}

\item{library_size}{Number of trancripts per cell/spot}

\item{spot_ct}{Number of cell types per spot}

\item{p}{Number of covariates}

\item{num_null}{Number of null coefficients to include}

\item{prob_ct}{Optional probabilities for cell type sampling}

\item{family}{The data generating distribution. One of poisson, negative binomial,binomial, and gaussian}

\item{dispersion}{Dispersion parameter for NB or Gaussian family}
}
\value{
A list with simulated y, X, lambda, beta, null_beta, and CT
}
\description{
This function simulates gene expression data using a spot-based model.
}
