% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sparrow}
\alias{sparrow}
\title{Provides basic information related to Statistical Analysis}
\usage{
sparrow(values)
}
\arguments{
\item{values}{A list of values.}
}
\value{
Returns the number of NA values in the \code{values}, the mean of the \code{values}, the range of the \code{values}, the quartile values of the \code{values}, the variance of the \code{values}, and the standard deviation of the \code{values}.
}
\description{
Provides basic information related to Statistical Analysis
}
\examples{
sparrow(c(2,3,4,5,7,8,9,10,13,15,17,21,22,23))
}
